+++ blueprints/ember-cli-page-object/index.js
@@ -1,8 +1,0 @@
-/*jshint node:true*/
-module.exports = {
-  normalizeEntityName: function() {},
-
-  afterInstall: function() {
-    return this.addBowerPackageToProject('ceibo', '1.1.0');
-  }
-};