+++ tests/unit/actions/click-on-text-test.js
@@ -139,6 +139,7 @@ test('resets scope', function(assert) {
 });
 
 test('returns target object', function(assert) {
+  fixture('<button>dummy text</button>');
   assert.expect(1);
 
   let page;
@@ -187,4 +188,74 @@ test('finds element by index', function(assert) {
     expectedSelector = 'button:contains("Lorem"):eq(2)';
     page.bar('Lorem');
   });
+});
+
+test('looks for elements outside the testing container', function(assert) {
+  assert.expect(1);
+
+  fixture('<button>Lorem</button>', { useAlternateContainer: true });
+
+  let expectedContext = '#alternate-ember-testing',
+      page;
+
+  window.click = function(_, actualContext) {
+    assert.equal(actualContext, expectedContext);
+  };
+
+  page = create({
+    foo: clickOnText('button', { testContainer: expectedContext })
+  });
+
+  andThen(() => {
+    page.foo('Lorem');
+  });
+});
+
+test("raises an error when the element doesn't exist", function(assert) {
+  assert.expect(1);
+
+  var done = assert.async();
+
+  let page = create({
+    foo: {
+      bar: {
+        baz: {
+          qux: clickOnText('button')
+        }
+      }
+    }
+  });
+
+  page.foo.bar.baz.qux('Lorem').then().catch(error => {
+    assert.ok(/page\.foo\.bar\.baz\.qux/.test(error.toString()), 'Element not found');
+  }).finally(done);
+});
+
+test("doesn't raise an error when the element is not visible and `visible` is not set", function(assert) {
+  fixture('<button style="display:none">Click me</button>');
+  assert.expect(1);
+
+  window.click = function() {
+    assert.ok(true, 'Element is clicked');
+  };
+
+  let page = create({
+    foo: clickOnText('button')
+  });
+
+  page.foo('Click me');
+});
+
+test('raises an error when the element is not visible and `visible` is true', function(assert) {
+  fixture('<button style="display:none">Click me</button>');
+  assert.expect(1);
+
+  let done = assert.async();
+  let page = create({
+    foo: clickOnText('button', { visible: true })
+  });
+
+  page.foo('Click me').then().catch(error => {
+    assert.ok(/page\.foo/.test(error.toString()), 'Element not found');
+  }).finally(done);
 });